Extract shared parts of cyclefold.
