+++ nova/src/folding/mod.rs
@@ -1,2 +1,3 @@
 pub mod hypernova;
 pub mod nova;
+pub(crate) mod cyclefold;
